+++ ikan/interface/editor/property_grid.hpp
@@ -85,8 +85,8 @@ namespace ikan {
                        float& value,
                        bool* checkbox_flag = nullptr,
                        float delta = 0.1f,
-                       int32_t min_value = INT_MIN,
-                       int32_t max_value = INT_MAX,
+                       float min_value = INT_MIN,
+                       float max_value = INT_MAX,
                        float reset_value = 0.0f,
                        float column_width = 100.0f);
 
@@ -104,8 +104,8 @@ namespace ikan {
                        glm::vec2& value,
                        bool* checkbox_flag = nullptr,
                        float delta = 0.1f,
-                       int32_t min_value = INT_MIN,
-                       int32_t max_value = INT_MAX,
+                       float min_value = INT_MIN,
+                       float max_value = INT_MAX,
                        float reset_value = 0.0f,
                        float column_width = 100.0f );
     
@@ -123,8 +123,8 @@ namespace ikan {
                        glm::vec3& value,
                        bool* checkbox_flag = nullptr,
                        float delta = 0.1f,
-                       int32_t min_value = INT_MIN,
-                       int32_t max_value = INT_MAX,
+                       float min_value = INT_MIN,
+                       float max_value = INT_MAX,
                        float reset_value = 0.0f,
                        float column_width = 100.0f );
 
@@ -225,8 +225,8 @@ namespace ikan {
                           const std::vector<float*>& values,
                           bool* checkbox_flag,
                           float delta,
-                          int32_t min_value,
-                          int32_t max_value,
+                          float min_value,
+                          float max_value,
                           float reset_value,
                           float column_width);
